% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsiter.R
\name{iterate}
\alias{iterate}
\title{Perform one iteration of segmentation. Updates by reference and
should be an internal function.}
\usage{
iterate(data, max_error)
}
\arguments{
\item{data}{data.table that has been setup by \code{setup}}

\item{max_error}{stopping criteria from \code{tdtr}}
}
\description{
Perform one iteration of segmentation. Updates by reference and
should be an internal function.
}
