% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{meanFilter}
\alias{meanFilter}
\title{Mean filter}
\usage{
meanFilter(coord, n = 3)
}
\arguments{
\item{coord}{A vector of coordinates over which to apply a mean filter}

\item{n}{The number of values to average}
}
\value{
A vector of mean-averaged coordinates
}
\description{
Mean filter
}
