% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{medianFilter}
\alias{medianFilter}
\title{Median filter}
\usage{
medianFilter(coord, n = 3)
}
\arguments{
\item{coord}{A vector of coordinates over which to apply a mean filter}

\item{n}{The number of values to average (best when odd-numbered)}
}
\value{
A vector of median-averaged coordinates
}
\description{
Median filter
}
