% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsiter.R
\name{setup}
\alias{setup}
\title{Set up a data.table for iterative segmentation}
\usage{
setup(data)
}
\arguments{
\item{data}{A data.frame or data.table containing lat, lon and timestamp}
}
\value{
A data.table with numeric timestamp, and an initial segment
}
\description{
Set up a data.table for iterative segmentation
}
