% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-calc.R
\name{singleSegmentSynchError}
\alias{singleSegmentSynchError}
\title{Calculate the Single Segment Synchronous Error on a \code{tdtr} object}
\usage{
singleSegmentSynchError(res, tol = 1e-24)
}
\arguments{
\item{res}{Segmented results from \code{tdtr}}

\item{tol}{Tolerance for zero comparisons. May vary with time scale.}
}
\value{
Adds by reference a column with segment error as given in Meratnia & de By \(2004\)
}
\description{
\code{singleSegmentSynchError} solves for the average distance with respect
to time between the approximated segment returned by the top-down time ratio
algorithm and the 'true' segment formed by two adjacent recorded points. This
is currently in development.
}
