% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{splitDiffTime}
\alias{splitDiffTime}
\title{Split the difference when it comes to difftimes}
\usage{
splitDiffTime(timestamp)
}
\arguments{
\item{timestamp}{a duration, period, difftime or interval}
}
\value{
the averaged difftime of same length
}
\description{
Averages out time differences between successive locations. This is useful
in calculating the time-weighted Radius of Gyration, as it provides a method
of using both the first and last locations. This assumes that the location
is measured at a given time period and will account for half of the time
difference occurring between this location and the one immediately preceding,
as well as half the time difference occurring between this location and the
one immediately following.
}
