% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{getSegments}
\alias{getSegments}
\title{Extract segment info from the segmented data.table}
\usage{
getSegments(data, coord.type = c("coordinate", "distance", "both"))
}
\arguments{
\item{data}{data.table returned from function /code{tdtr}}

\item{coord.type}{return actual coordinates, relative distance, or both}
}
\value{
data.table with segments only
}
\description{
Extract segment info from the segmented data.table
}
