% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{getSegsExtra}
\alias{getSegsExtra}
\title{Extract segment info with extra information from the segmented data.table}
\usage{
getSegsExtra(
  data,
  coord.type = c("coordinate", "distance", "both"),
  group = FALSE
)
}
\arguments{
\item{data}{data.table returned from function /code{tdtr}}

\item{coord.type}{return actual coordinates, relative distance, or both}

\item{group}{Separate by group, default is FALSE}
}
\value{
data.table with segments only
}
\description{
Extract segment info with extra information from the segmented data.table
}
