% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsiter.R
\name{iterate}
\alias{iterate}
\title{Perform one iteration of segmentation}
\usage{
iterate(data)
}
\arguments{
\item{data}{data.table that has been setup by \code{setup}}
}
\value{
data.table with one additional segment and error distance
}
\description{
Perform one iteration of segmentation
}
