% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsiter.R
\name{tdtr}
\alias{tdtr}
\title{Perform Top-Down Time Ratio segmentation}
\usage{
tdtr(
  data,
  col_names = list(timestamp_col = "timestamp", latitude_col = "lat", longitude_col =
    "lon"),
  max_segs = 5000,
  n_segs = max_segs,
  max_error = 200
)
}
\arguments{
\item{data}{is a data.frame or data.table with timestamp, lat and lon}

\item{col_names}{named list with column names for timestamp, lat and lon}

\item{max_segs}{with maximum number of segments allowed default 5000}

\item{n_segs}{used to generate a specific number of segments}

\item{max_error}{used as stopping criteria}
}
\value{
data.table with segment information
}
\description{
Perform Top-Down Time Ratio segmentation
}
