% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topdowntimeratio-package.R
\docType{package}
\name{topdowntimeratio-package}
\alias{topdowntimeratio}
\alias{topdowntimeratio-package}
\title{topdowntimeratio: Top-Down Time Ratio segmentation for coordinate trajectories}
\description{
This package applies the compression technique specified in Meratnia and de By (2004) \doi{10.1007/978-3-540-24741-8} on timestamped coordinate trajectories. This is an improvement on the well-known Douglas-Peucker algorithm for segmentation that operates not on the basis of perpendicular distances, but instead distances with respect to time. Provided a trajectory with timestamps, tdtr() returns a set of straight- line segments that can represent the full trajectory. McCool, Lugtig, and Schouten (2022) \doi{10.1007/s11116-022-10328-2} describe this method as implemented here in more detail.
}
\author{
\strong{Maintainer}: Danielle McCool \email{d.m.mccool@uu.nl} (\href{https://orcid.org/0000-0002-7055-7539}{ORCID})

}
\keyword{internal}
